requirements.txt
Flask
numpy
Pillow
tensorflow
Werkzeug
